+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
+## [v4.3.1] - 2020-08-25
+- Enforce controller API consistency
+- Activate module after installation
+
 ## [v4.3.0] - 2020-07-16
 - Utilize timezones
 - Use app('request.user') instead of $currentUser
@@ -155,6 +159,7 @@ This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 ## v0.0.1 - 2017-09-09
 - Tag first release
 
+[v4.3.1]: https://github.com/rinvex/cortex-categories/compare/v4.3.0...v4.3.1
 [v4.3.0]: https://github.com/rinvex/cortex-categories/compare/v4.2.2...v4.3.0
 [v4.2.2]: https://github.com/rinvex/cortex-categories/compare/v4.2.1...v4.2.2
 [v4.2.1]: https://github.com/rinvex/cortex-categories/compare/v4.2.0...v4.2.1